vision/requirements.txt
mediapipe>=0.10.9
opencv-python>=4.8.0
numpy>=1.24.0
pyserial>=3.5
